Cambios generales estructurales

+++ Models/OrdenCompra.cs
@@ -2,7 +2,7 @@
 {
     public class OrdenCompra
     {
-        public virtual int nOrdenCompra {  get; set; }
+        public virtual long nOrdenCompra {  get; set; }
         public virtual string? detalleOrdenCompra {  set; get; }
         public virtual long totalPagar { get; set; }
         public virtual OrdenCompraEstado? ordenEstado { get; set; }
