+++ Models/StockArticulos.cs
@@ -3,8 +3,8 @@
     public class StockArticulos
     {
         public virtual long nStock { get; set; }
-        public virtual int stockSeguridad {  get; set; }
-        public virtual int stockActual {  get; set; }
+        public virtual long stockSeguridad {  get; set; }
+        public virtual long stockActual {  get; set; }
         public virtual DateTime fechaStockInicio { get; set; }
         public virtual DateTime? fechaStockFin { get; set; }
         public virtual Articulo? articulo { get; set; }
